+++ src/commands/copyCurrentBranch.ts
@@ -1,13 +1,12 @@
 'use strict';
-
-import { env , TextEditor , Uri , window } from 'vscode';
-import { GitService } from '../git/gitService';
+import { env, TextEditor, Uri, window } from 'vscode';
+import { ActiveEditorCommand, command, Commands, getCommandUri, getRepoPathOrActiveOrPrompt } from './common';
+import { Container } from '../container';
 import { GitUri } from '../git/gitUri';
 import { Logger } from '../logger';
-import { ActiveEditorCommand, command, Commands, getCommandUri, getRepoPathOrActiveOrPrompt } from './common';
 
 @command()
-export class CopyCurrentBranch extends ActiveEditorCommand {
+export class CopyCurrentBranchCommand extends ActiveEditorCommand {
 	constructor() {
 		super(Commands.CopyCurrentBranch);
 	}
@@ -17,17 +16,17 @@ export class CopyCurrentBranch extends ActiveEditorCommand {
 
 		const gitUri = uri != null ? await GitUri.fromUri(uri) : undefined;
 
-		const repoPath = await getRepoPathOrActiveOrPrompt(gitUri, editor, 'Copy Current Branch');
+		const repoPath = await getRepoPathOrActiveOrPrompt(gitUri, editor, 'Copy Current Branch Name');
 		if (!repoPath) return;
 
-		const service = new GitService();
-
 		try {
-			const gitBranch = await service.getBranch(repoPath);
-			if (gitBranch?.name) await env.clipboard.writeText(gitBranch?.name);
+			const branch = await Container.git.getBranch(repoPath);
+			if (branch?.name) {
+				await env.clipboard.writeText(branch.name);
+			}
 		} catch (ex) {
-			Logger.error(ex, 'CopyCurrentBranch');
-			void window.showErrorMessage('Unable to copy current branch. See output channel for more details');
+			Logger.error(ex, 'CopyCurrentBranchCommand');
+			void window.showErrorMessage('Unable to copy current branch name. See output channel for more details');
 		}
 	}
 }